LayoutTests/imported/w3c/web-platform-tests/fetch/private-network-access/nested-worker.https.window-expected.txt


FAIL treat-as-public to local: failure. assert_equals: worker error expected (string) "unknown error" but got (undefined) undefined
FAIL treat-as-public to private: failure. assert_equals: worker error expected (string) "unknown error" but got (undefined) undefined
PASS public to public: success.
